whereismybus/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.36
pydantic>=2.9.0
bcrypt>=4.2.0
pyjwt>=2.9.0
python-multipart>=0.0.12
httpx>=0.27.0
